2860344  ooDialog - unify dialog control methods
To instantiate a dialog control object the method names will be newXXX() with
XXX being the unified control name.  I.e. edit, treeView, listView, listBox,
etc..

To connect a dialog control to a dialog attribute and add it to the data
table, the method names will be connectXXX() with XXX beind the unified
control name.

The set and get 'data' methods will also use unified method names.  The names
will contain the word data in them to distinguish them from other set and get
methods names.  I.e., setXXXData() or getXXXData() where XXX is the unified
control name.  These methods have been reworked to remove the use of interpret
from them on the Rexx side and to use an enum to identify the type of dialog
control, rathar than using magic numbers, on the C / C++ side.

The use of deprecated methods in the ooDialog sample programs, related to the
above changes, has been fixed.



git-svn-id: svn+ssh://svn.code.sf.net/p/oorexx/code-0/main/trunk@5217 0b6cbdbe-3aab-466e-b73a-abd511dda0a2