ooDialog -
Fix up the DlgAreaU class so that the user can control whether
the dialog is forced to update on every resize event.

Add the connectSizeMoveEnded method to the event notification class. Fix up
the event handler invocation code so that, with some connect methods, the user
can specify how the event handler is invoked, directly or through startWith().

Finish up the setResizing method and rename it to diasableInternalResize
method of the Edit class.



git-svn-id: svn+ssh://svn.code.sf.net/p/oorexx/code-0/main/trunk@6565 0b6cbdbe-3aab-466e-b73a-abd511dda0a2