ooDialog - For the dlgAreaUDemoTwo example:
* Under some conditions, the dialog was not repainting correctly.  The
DlgAreaU class is actually using the wrong method to force the repaint.  The
update() method invalidates the client area, but does not force the window
to immediately redraw.  The proper method is redrawClient()

* Since in the example, we are not redrawing until the user stops resizing,
there is no point in recalculating the size and position on every resize
event.  Instead, just do that once, when the user has stopped sizing the
dialog.


git-svn-id: svn+ssh://svn.code.sf.net/p/oorexx/code-0/main/trunk@6598 0b6cbdbe-3aab-466e-b73a-abd511dda0a2